rename root_task to explicit_task
