+++ src/search/tasks/explicit_task.cc
@@ -1,4 +1,4 @@
-#include "root_task.h"
+#include "explicit_task.h"
 
 #include "../axioms.h"
